implemented func Getvisiter, added additional method to Human

+++ eatable.go
@@ -2,5 +2,5 @@ package hrw
 
 type Eatable struct {
 	Name     string
-	Calories uint16
+	Calories uint8
 }
